Benchmark for File#each_byte.
This benchmark was added for https://github.com/rubinius/rubinius/issues/2166
to make it easier to run and keep track of. It does also benchmark File#rewind
but I'm not sure if there's a way around that. The alternative would be to use
File.open() for each iteration but then we'd end up benchmarking even more.